requirements.txt
pyyaml==6.0
pytest-playwright==0.3.0
rich==12.6.0
Flask==2.2.2
